Add common::fs module * Move common::file_lock to common::fs::file_lock * Move slurp, safe_slurp, file_lock, remove_file, write, read_dir_async, read_dir, json_dir_entries to common::fs
