+++ mm2src/mm2_tests.rs
@@ -23,7 +23,7 @@ use uuid::Uuid;
 
 cfg_native! {
     use common::for_tests::{get_passphrase, new_mm2_temp_folder_path};
-    use common::slurp;
+    use common::fs::slurp;
     use hyper::header::ACCESS_CONTROL_ALLOW_ORIGIN;
 }
